linux-intel: add LZ4 dependency
A change was made on upstream, to move from LZ4C to LZ4. This makes sense
on master branch, but on Scarthgap we end up with a build issue.

To fix this for Scarthgap, we add the 'lz4-native' dependency on the
recipe.

[1] https://github.com/intel/linux-intel-lts/commit/cbfb30ae17d45b92bc9e5fb119a4a6a81433c5cb

Related-to: TOR-4064

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>
(cherry picked from commit dda6c83ba5bea0419ed5118628941a1b43dead8a)